requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.5
aiofiles>=0.6.0
pillow>=8.0.0
python-dotenv>=0.19.0
fastapi-mail>=1.2.0
email-validator>=1.1.3
jinja2>=3.0.1
httpx>=0.24.0  # Added for UploadThing integration
